requirements.txt
streamlit
openai-whisper
transformers
torch
pandas
Pillow
streamlit-audiorecorder
sentence-transformers
keybert
scikit-learn
setuptools